+++ l10n_ar_fiscal_ws_fe/__manifest__.py
@@ -6,19 +6,20 @@
     "license": "AGPL-3",
     "summary": "",
     "depends": [
-        "l10n_ar_afipws",
+        "l10n_ar_fiscal_ws",
         "l10n_ar",
         "account_debit_note",
     ],
     "external_dependencies": {},
     "data": [
-        "views/account_move_views.xml",
+        "data/arcaws.xml",
+        #"views/account_move_views.xml",
         "views/account_journal_view.xml",
-        "views/report_invoice.xml",
-        "views/res_config_settings.xml",
+        #"views/report_invoice.xml",
+        #"views/res_config_settings.xml",
         "views/menuitem.xml",
-        "views/ir_cron.xml",
-        "wizard/account_validate_account_move.xml",
+        #"views/ir_cron.xml",
+        #"wizard/account_validate_account_move.xml",
     ],
     "demo": [],
     "images": [],
