+++ l10n_ar_fiscal_ws_fe/models/__init__.py
@@ -2,5 +2,9 @@
 # For copyright and license notices, see __manifest__.py file in module root
 # directory
 ##############################################################################
-from . import models
-from . import wizard
+from . import account_move
+from . import account_move_ws
+from . import account_journal
+#from . import account_journal_ws
+from . import res_config_settings
+from . import res_company
