+++ l10n_ar_fiscal_ws_fe/views/account_journal_view.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="view_account_journal_form" model="ir.ui.view">
+        <field name="name">account.journal.form</field>
+        <field name="model">account.journal</field>
+        <field name="inherit_id" ref="account.view_account_journal_form"/>
+        <field name="arch" type="xml">
+            <sheet position="before">
+                <header>
+                    <button name="action_get_connection" string="Get Connection" help="Get Connection For this webservice and create it if no valid" type="object" invisible="not arcaws"/>
+                    <button name="test_pyafipws_dummy" string="Dummy Test" help="Dummy test to check connection to selected webservice" type="object" invisible="not arcaws"/>
+                    <button name="test_pyafipws_point_of_sales" string="Get Points Of Sale" help="Get Enable Point of Sales for this webservice" type="object" invisible="not arcaws"/>
+                    <button name="get_pyafipws_cuit_document_classes" string="Get Document Types" help="Get valid document types for this webservice" type="object" invisible="not arcaws"/>
+                    <!--<button name="get_pyafipws_zonas" string="Get Zones" help="Get zones for this webservice" type="object" invisible="arcaws != 'wsbfe'"/>
+                    <button name="get_pyafipws_NCM" string="Get NCM" help="Obtener códigos del Nomenclador Común del Mercosur" type="object" invisible="arcaws != 'wsbfe'"/>
+                    <button name="get_pyafipws_post_invoice_numbers" string="Get next Inv" help="Obtener numeros de factura" type="object" invisible="not arcaws"/> -->
+                </header>
+            </sheet>
+        </field>
+    </record>
+</odoo>
