+++ l10n_ar_fiscal_ws_fe/views/res_config_settings.xml
